Move State and Event to have only the events implemented in the state machine in the states folder
