+++ uiWebRobot/state_machine/states/CreateFieldState.py
@@ -6,10 +6,10 @@ import navigation
 import posix_ipc
 import threading
 
-from state_machine.states import State
+from state_machine import State
 from state_machine.states import WaitWorkingState
 from state_machine.states import ErrorState
-from state_machine.states import Events
+from state_machine import Events
 from state_machine.FrontEndObjects import FrontEndObjects, ButtonState
 from state_machine.utilsFunction import *
 from config import config
