+++ uiWebRobot/state_machine/states/WaitWorkingState.py
@@ -4,12 +4,12 @@ sys.path.append('../')
 from flask_socketio import SocketIO
 import threading
 
-from state_machine.states import State
+from state_machine import State
 from state_machine.states import CreateFieldState
 from state_machine.states import StartingState
 from state_machine.states import ResumeState
 from state_machine.states import ErrorState
-from state_machine.states import Events
+from state_machine import Events
 
 from state_machine.FrontEndObjects import FrontEndObjects, ButtonState, AuditButtonState
 from state_machine.utilsFunction import *
