+++ uiWebRobot/state_machine/states/WorkingState.py
@@ -7,10 +7,10 @@ import posix_ipc
 import threading
 from datetime import datetime, timezone
 
-from state_machine.states import State
+from state_machine import State
 from state_machine.states import WaitWorkingState
 from state_machine.states import ErrorState
-from state_machine.states import Events
+from state_machine import Events
 
 from uiWebRobot.state_machine.FrontEndObjects import AuditButtonState, ButtonState, FrontEndObjects
 from uiWebRobot.state_machine.utilsFunction import *
